scraper/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
praw
openai
supabase
